+++ lib/src/measurements/derived.dart
@@ -9,53 +9,61 @@ class DerivedMeasurement<A extends Measurement<A>, B extends Measurement<B>>
     extends Measurement<DerivedMeasurement<A, B>> {
   /// The derived measurement of zero magnitude.
   const DerivedMeasurement.zero(
-      final DerivedMeasurementInterpreter<A, B> interpreter)
+      [final DerivedMeasurementInterpreter<A, B>? interpreter])
       : _reciprocal = false,
         super.zero(interpreter);
 
   /// The derived measurement of infinite magnitude.
   const DerivedMeasurement.infinite(
-      final DerivedMeasurementInterpreter<A, B> interpreter)
+      [final DerivedMeasurementInterpreter<A, B>? interpreter])
       : _reciprocal = false,
         super.infinite(interpreter);
 
   /// The derived measurement of negative infinite magnitude.
   const DerivedMeasurement.negativeInfinite(
-      final DerivedMeasurementInterpreter<A, B> interpreter)
+      [final DerivedMeasurementInterpreter<A, B>? interpreter])
       : _reciprocal = false,
         super.negativeInfinite(interpreter);
 
   /// Constructs a derived measurement representing the multiplication of a pair
   /// of simpler measurements.
   DerivedMeasurement.multiply(final A a, final B b, [final String? name])
-      : this._(
-          a.si * b.si,
-          Precision.combine(a._precision, b._precision),
-          DerivedMeasurementInterpreter(
+      : this._combine(
+            a.si * b.si,
+            Precision.combine(a._precision, b._precision),
             a.defaultInterpreter,
             b.defaultInterpreter,
             false,
-            const MeasurementPrefix.unit(),
-            name,
-          ),
-          false,
-        );
+            name);
 
   /// Constructs a derived measurement representing the division of a pair of
   /// simpler measurements.
   DerivedMeasurement.divide(final A a, final B b, [final String? name])
-      : this._(
-          a.si / b.si,
-          Precision.combine(a._precision, b._precision),
-          DerivedMeasurementInterpreter(
+      : this._combine(
+            a.si / b.si,
+            Precision.combine(a._precision, b._precision),
             a.defaultInterpreter,
             b.defaultInterpreter,
             true,
-            const MeasurementPrefix.unit(),
-            name,
-          ),
-          true,
-        );
+            name);
+
+  /// Creates a derived measurement representing a pair of simpler units.
+  DerivedMeasurement._combine(double value, Precision precision,
+      [final MeasurementInterpreter<A>? a,
+      final MeasurementInterpreter<B>? b,
+      final bool reciprocal = false,
+      final String? name])
+      : this._(
+            value,
+            precision,
+            DerivedMeasurementInterpreter._unitless(
+              a,
+              b,
+              reciprocal,
+              const MeasurementPrefix.unit(),
+              name,
+            ),
+            reciprocal);
 
   /// Interprets this using two specific units.
   double as(final MeasurementInterpreter<A> a,
@@ -74,16 +82,16 @@ class DerivedMeasurement<A extends Measurement<A>, B extends Measurement<B>>
   @override
   DerivedMeasurement<A, B> _construct(
     final double si,
-    final Precision precision,
-    final MeasurementInterpreter<DerivedMeasurement<A, B>> interpreter,
-  ) =>
+    final Precision precision, [
+    final MeasurementInterpreter<DerivedMeasurement<A, B>>? interpreter,
+  ]) =>
       DerivedMeasurement._(si, precision, interpreter);
 
   /// Constructs a derived measurement.
   const DerivedMeasurement._(
     final double si,
-    final Precision precision,
-    final MeasurementInterpreter<DerivedMeasurement<A, B>> interpreter, [
+    final Precision precision, [
+    final MeasurementInterpreter<DerivedMeasurement<A, B>>? interpreter,
     this._reciprocal = false,
   ]) : super(si, precision, interpreter);
 
@@ -102,10 +110,10 @@ class DerivedMeasurementInterpreter<A extends Measurement<A>,
   final bool _reciprocal;
 
   /// The numerator's interpreter.
-  final MeasurementInterpreter<A> _interpreterA;
+  final MeasurementInterpreter<A>? _interpreterA;
 
   /// The denominator's interpreter.
-  final MeasurementInterpreter<B> _interpreterB;
+  final MeasurementInterpreter<B>? _interpreterB;
 
   /// Constructs an interpreter from two basic interpreters.
   ///
@@ -118,23 +126,36 @@ class DerivedMeasurementInterpreter<A extends Measurement<A>,
     final bool reciprocal = false,
     final MeasurementPrefix prefix = const MeasurementPrefix.unit(),
     final String? name,
+  ]) : this._unitless(a, b, reciprocal, prefix, name);
+
+  /// Constructs a [DerivedMeasurementInterpreter] with a potentially null
+  /// interpreter.
+  DerivedMeasurementInterpreter._unitless([
+    final MeasurementInterpreter<A>? a,
+    final MeasurementInterpreter<B>? b,
+    final bool reciprocal = false,
+    final MeasurementPrefix prefix = const MeasurementPrefix.unit(),
+    final String? name,
   ])  : _reciprocal = reciprocal,
         _interpreterA = a,
         _interpreterB = b,
         super._(
-          name ?? '$a${reciprocal ? '/' : '⋅'}$b',
+          name ?? '${a ?? 'X'}${reciprocal ? '/' : '⋅'}${b ?? 'X'}',
           reciprocal
-              ? a._unitMultiplier / b._unitMultiplier
-              : a._unitMultiplier * b._unitMultiplier,
+              ? (a?._unitMultiplier ?? 1.0) / (b?._unitMultiplier ?? 1.0)
+              : (a?._unitMultiplier ?? 1.0) * (b?._unitMultiplier ?? 1.0),
           prefix,
         );
 
   @override
   DerivedMeasurement<A, B> call(final double value,
           {final Precision precision = Precision.max}) =>
-      _reciprocal
-          ? DerivedMeasurement.divide(
-              _interpreterA.call(value), _interpreterB.call(value))
-          : DerivedMeasurement.multiply(
-              _interpreterA.call(value), _interpreterB.call(value));
+      DerivedMeasurement._combine(
+        value,
+        precision,
+        _interpreterA,
+        _interpreterB,
+        _reciprocal,
+        _name,
+      );
 }