+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 Dependency-free, type-safe Dart unit conversion library. Decouple specific unit
 types you need to work with from the rest of your code.
 
+## [Unreleased]
+### Changed
+- Make default interpreters optional for all units
+
 ## [2.2.4]
 ### Added
 - Angle boolean checks: isAcute, isRight, isObtuse, isStraight